requirements.txt
python-dotenv
transformers==4.27.3 
langchain 
openai
requests
streamlit